codigo/raspi/requirements.txt
matplotlib>=3.7.1
numpy>=1.26.1
pandas<=2.2.2
paramiko==2.9.3
pyzmq==24.0.1
river==0.19.0
scikit-learn>=1.3.0
scipy>=1.11.1
screeninfo==0.8.1
simpy==4.1.1
tqdm==4.65.0
